refactor: more preps for mvc and better work with local
