+++ src/front/style.css
@@ -5,3 +5,7 @@
 .img {
   width: 100%;
 }
+
+.active {
+  border: 2px solid rgba(0, 64, 255, 0.003);
+}
